Implemented full CRUD functionality for the Event entity

+++ MyStagePass.Model/Models/Purchase.cs
@@ -5,15 +5,15 @@ using System.Linq;
 
 namespace MyStagePass.Model.Models
 {
-    public class Purchase
-    {
-        public int PurchaseID { get; set; }
-        public DateTime PurchaseDate { get; set; }
-        public int CustomerID { get; set; }
-       // public virtual Customer Customer { get; set; } = null!;
-        public virtual ICollection<Ticket> Tickets { get; set; } = new List<Ticket>();
+	public class Purchase
+	{
+		public int PurchaseID { get; set; }
+		public DateTime PurchaseDate { get; set; }
+		public int CustomerID { get; set; }
+		//public virtual Customer Customer { get; set; } = null!;
+		public virtual ICollection<Ticket> Tickets { get; set; } = new List<Ticket>();
 
-        [NotMapped]
-        public float Total => Tickets.Sum(i => i.Price);
-    }
+		[NotMapped]
+		public int Total => Tickets.Sum(i => i.Price);
+	}
 }
